Revert "RawInput方式で打鍵監視するやつ"
