+++ VMagicMirror/Assets/Baku/VMagicMirror/Prefabs/Monitoring/MonitoringLogics.prefab
@@ -50,8 +50,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: cc468fe23300d614bb995f234222176e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  globalHookInputChecker: {fileID: 3968200917505937623}
-  robustRawInputChecker: {fileID: 1881006542137438839}
+  rawInputChecker: {fileID: 3968200917505937623}
   mousePositionProvider: {fileID: 5115481182964325839}
   faceTracker: {fileID: 464634160456065510}
   handTracker: {fileID: 3822874109225115685}
@@ -275,18 +274,6 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: fdce11c20059ecd4dad29b0fc634b169, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!114 &1881006542137438839 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 5445794628714955643, guid: ee92b6c9100e05d4d88246a79d8af1d9,
-    type: 3}
-  m_PrefabInstance: {fileID: 5875495639101518092}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0f365c46d8ecbff4695e9ac1b5cb469c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1001 &6576671094749192196
