--- VMagicMirror/Assets/Baku/VMagicMirror/Scripts/InputMonitoring/KeyAndMouseInput/GlobalHookInputChecker.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b157506b0989896458c4443037aeea30
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
