--- VMagicMirror/Assets/Baku/VMagicMirror/Scripts/InputMonitoring/KeyAndMouseInput/IKeyboardEventSource.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ec3353151006398458f21bbc36d863b5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
