+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/InputMonitoring/KeyAndMouseInput/RawInputChecker.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0f365c46d8ecbff4695e9ac1b5cb469c
+guid: b157506b0989896458c4443037aeea30
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
