+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/InputMonitoring/KeyAndMouseInput/RawMouseMoveChecker.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fcdcad91c6d624c48ad2e4bba82ece1a
+guid: 0f365c46d8ecbff4695e9ac1b5cb469c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
