+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/Installer/MonitoringInstaller.cs
@@ -6,8 +6,7 @@ namespace Baku.VMagicMirror.Installer
 {
     public class MonitoringInstaller : InstallerBase
     {
-        [SerializeField] private GlobalHookInputChecker globalHookInputChecker = null;
-        [SerializeField] private RawInputChecker robustRawInputChecker = null;
+        [SerializeField] private RawInputChecker rawInputChecker = null;
         [SerializeField] private MousePositionProvider mousePositionProvider = null;
         [SerializeField] private FaceTracker faceTracker = null;
         [SerializeField] private HandTracker handTracker = null;
@@ -18,10 +17,7 @@ namespace Baku.VMagicMirror.Installer
         
         public override void Install(DiContainer container)
         {
-            //NOTE: 2つの実装が合体したキメラ実装を適用します。コレが比較的安全でいちばん動きも良いので。
-            container.Bind<IKeyMouseEventSource>()
-                .FromInstance(new HybridInputChecker(robustRawInputChecker, globalHookInputChecker))//globalHookInputChecker)
-                .AsCached();
+            container.BindInstance(rawInputChecker);
             container.BindInstance(mousePositionProvider);
             container.BindInstance(faceTracker);
             container.BindInstance(handTracker);
@@ -32,11 +28,10 @@ namespace Baku.VMagicMirror.Installer
 
             //終了前に監視処理を安全にストップさせたいものは呼んでおく
             container.Bind<IReleaseBeforeQuit>()
-                .FromInstance(robustRawInputChecker)
+                .FromInstance(mousePositionProvider)
                 .AsCached();
-
             container.Bind<IReleaseBeforeQuit>()
-                .FromInstance(globalHookInputChecker)
+                .FromInstance(rawInputChecker)
                 .AsCached();
         }
     }
